Testing another encoder that doesn't call functions from flash in ISR
On ESP12, it is not allowed to call flash functions from an interrupt
context. This means digitalRead() are not allowed. This new version of
the encoder relies entirely on interrupts and variables to keep track
of GPIO values. The callbacks are also removed since the user may not
know about having his functions in RAM.